requirements.txt
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
jwt==0.6.1
Pillow==6.2.1
pycparser==2.19
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
